[mypyc] Refactor: extract RType related code from mypyc.ops
The only trickiness was dealing with cyclic imports that this
generated.
