+++ mypyc/emit.py
@@ -7,9 +7,9 @@ from mypyc.common import (
     REG_PREFIX, ATTR_PREFIX, STATIC_PREFIX, TYPE_PREFIX, NATIVE_PREFIX,
     FAST_ISINSTANCE_MAX_SUBCLASSES,
 )
-from mypyc.ops import (
-    Environment, BasicBlock, Value, RType, RTuple, RInstance,
-    RUnion, RPrimitive,
+from mypyc.ops import Environment, BasicBlock, Value
+from mypyc.rtypes import (
+    RType, RTuple, RInstance, RUnion, RPrimitive,
     is_float_rprimitive, is_bool_rprimitive, is_int_rprimitive, is_short_int_rprimitive,
     is_list_rprimitive, is_dict_rprimitive, is_set_rprimitive, is_tuple_rprimitive,
     is_none_rprimitive, is_object_rprimitive, object_rprimitive, is_str_rprimitive,
