+++ mypyc/emitfunc.py
@@ -9,10 +9,10 @@ from mypyc.emit import Emitter
 from mypyc.ops import (
     OpVisitor, Goto, Branch, Return, Assign, LoadInt, LoadErrorValue, GetAttr, SetAttr,
     LoadStatic, InitStatic, TupleGet, TupleSet, Call, IncRef, DecRef, Box, Cast, Unbox,
-    BasicBlock, Value, RType, RTuple, MethodCall, PrimitiveOp,
-    EmitterInterface, Unreachable, NAMESPACE_STATIC, NAMESPACE_TYPE, NAMESPACE_MODULE,
-    RaiseStandardError
+    BasicBlock, Value, MethodCall, PrimitiveOp, EmitterInterface, Unreachable, NAMESPACE_STATIC,
+    NAMESPACE_TYPE, NAMESPACE_MODULE, RaiseStandardError
 )
+from mypyc.rtypes import RType, RTuple
 from mypyc.func_ir import FuncIR, FuncDecl, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
 from mypyc.class_ir import ClassIR
 from mypyc.namegen import NameGenerator
