+++ mypyc/emitwrapper.py
@@ -6,7 +6,7 @@ from mypy.nodes import ARG_POS, ARG_OPT, ARG_NAMED_OPT, ARG_NAMED, ARG_STAR, ARG
 
 from mypyc.common import PREFIX, NATIVE_PREFIX, DUNDER_PREFIX
 from mypyc.emit import Emitter
-from mypyc.ops import (
+from mypyc.rtypes import (
     RType, is_object_rprimitive, is_int_rprimitive, is_bool_rprimitive, object_rprimitive
 )
 from mypyc.func_ir import FuncIR, RuntimeArg, FUNC_STATICMETHOD
