+++ mypyc/func_ir.py
@@ -3,10 +3,11 @@ from typing_extensions import Final
 
 from mypy.nodes import FuncDef, Block, ARG_POS, ARG_OPT, ARG_NAMED_OPT
 
+from mypyc.common import JsonDict
 from mypyc.ops import (
-    RType, JsonDict, DeserMaps, Goto, Branch, Return, Unreachable, BasicBlock, Environment,
-    deserialize_type
+    DeserMaps, Goto, Branch, Return, Unreachable, BasicBlock, Environment
 )
+from mypyc.rtypes import RType, deserialize_type
 from mypyc.namegen import NameGenerator
