+++ mypyc/irbuild/prepare.py
@@ -7,9 +7,8 @@ from mypy.nodes import (
 from mypy.types import Type
 from mypy.build import Graph
 
-from mypyc.ops import (
-    RInstance, tuple_rprimitive, dict_rprimitive, DeserMaps
-)
+from mypyc.ops import DeserMaps
+from mypyc.rtypes import RInstance, tuple_rprimitive, dict_rprimitive
 from mypyc.func_ir import (
     FuncDecl, FuncSignature, RuntimeArg, FUNC_NORMAL, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
 )
