+++ mypyc/irbuild/util.py
@@ -5,7 +5,9 @@ from mypy.nodes import (
     IntExpr, FloatExpr, Var, TupleExpr, UnaryExpr, BytesExpr, ARG_NAMED, ARG_NAMED_OPT, ARG_POS,
     ARG_OPT, GDEF
 )
-from mypyc.ops import Environment, RInstance, AssignmentTargetRegister
+
+from mypyc.ops import Environment, AssignmentTargetRegister
+from mypyc.rtypes import RInstance
 from mypyc.class_ir import ClassIR
 from mypyc.common import SELF_NAME
