+++ mypyc/primitives/dict_ops.py
@@ -2,11 +2,9 @@
 
 from typing import List
 
-from mypyc.ops import (
-    EmitterInterface,
-    dict_rprimitive, object_rprimitive, bool_rprimitive, int_rprimitive,
-    ERR_FALSE, ERR_MAGIC, ERR_NEVER,
-)
+from mypyc.ops import EmitterInterface, ERR_FALSE, ERR_MAGIC, ERR_NEVER
+from mypyc.rtypes import dict_rprimitive, object_rprimitive, bool_rprimitive, int_rprimitive
+
 from mypyc.primitives.registry import (
     name_ref_op, method_op, binary_op, func_op, custom_op,
     simple_emit, negative_int_emit, call_emit, call_negative_bool_emit,
