+++ mypyc/primitives/exc_ops.py
@@ -1,10 +1,7 @@
 """Exception-related primitive ops."""
 
-
-from mypyc.ops import (
-    bool_rprimitive, object_rprimitive, void_rtype, exc_rtuple,
-    ERR_NEVER, ERR_FALSE
-)
+from mypyc.ops import ERR_NEVER, ERR_FALSE
+from mypyc.rtypes import bool_rprimitive, object_rprimitive, void_rtype, exc_rtuple
 from mypyc.primitives.registry import (
     simple_emit, custom_op,
 )
