+++ mypyc/primitives/list_ops.py
@@ -2,10 +2,9 @@
 
 from typing import List
 
-from mypyc.ops import (
-    int_rprimitive, short_int_rprimitive, list_rprimitive, object_rprimitive, bool_rprimitive,
-    ERR_MAGIC, ERR_NEVER,
-    ERR_FALSE, EmitterInterface,
+from mypyc.ops import ERR_MAGIC, ERR_NEVER, ERR_FALSE, EmitterInterface
+from mypyc.rtypes import (
+    int_rprimitive, short_int_rprimitive, list_rprimitive, object_rprimitive, bool_rprimitive
 )
 from mypyc.primitives.registry import (
     name_ref_op, binary_op, func_op, method_op, custom_op, simple_emit,
