+++ mypyc/primitives/misc_ops.py
@@ -1,10 +1,10 @@
 """Miscellaneous primitive ops."""
 
 
-from mypyc.ops import (
+from mypyc.ops import ERR_NEVER, ERR_MAGIC, ERR_FALSE
+from mypyc.rtypes import (
     RTuple, none_rprimitive, bool_rprimitive, object_rprimitive, str_rprimitive,
-    int_rprimitive, dict_rprimitive,
-    ERR_NEVER, ERR_MAGIC, ERR_FALSE
+    int_rprimitive, dict_rprimitive
 )
 from mypyc.primitives.registry import (
     name_ref_op, simple_emit, binary_op, unary_op, func_op, method_op, custom_op,
