+++ mypyc/sametype.py
@@ -1,6 +1,6 @@
 """Same type check for RTypes."""
 
-from mypyc.ops import (
+from mypyc.rtypes import (
     RType, RTypeVisitor, RInstance, RPrimitive, RTuple, RVoid, RUnion
 )
 from mypyc.func_ir import FuncSignature
